+++ php/tests/Application/Command/VariableTypesTest/FunctionParameterTypeHintSpecial.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace A;
+
+class C
+{
+    /**
+     * @param static $b
+     * @param $this  $c
+     */
+    public function foo(self $a, $b, $c)
+    {
+        // <MARKER>
+    }
+}